WIP: reconciled sound-engine, format and DB work (mega-commit, not yet split)
Full pre-split snapshot of the sound-engine refactor branch, reconciled after
an incomplete cherry-pick from SuperView had dropped 39 files (ByteParsing's
DoubleParser and friends, the DB table parsing strand, UnitAudioSwitchResolver,
HircGraphService, scene-object work, localisation) and left SuperViewViewModel
using fields it never declared. Recovered the missing content from SuperView's
tip and from a stash of the same generation via per-file three-way merges,
keeping this branch's newer Phase 20 version-definition work and the newer
pack-file/encryption/settings work intact throughout.

This is one commit covering everything so the reconciled state is safely
stored before it gets split into the separate strands described in
Research/AiDocumentation/mega-commit-untanglement.md (Wwise format fixes,
Wwise format additions, DB table parsing, the audio engine, Super View/
animation, and other tooling). Not intended to stay as one commit.

Builds clean; Test.Audio 267 passed, Shared.CoreTest 397 passed/6 skipped.

+++ Shared/GameFiles/Wwise/Hirc/V136/Shared/MusicNodeParams_V136.cs
@@ -34,8 +34,8 @@ namespace Shared.GameFormats.Wwise.Hirc.V136.Shared
 
             public void ReadData(ByteChunk chunk)
             {
-                GridPeriod = chunk.ReadInt64();
-                GridOffset = chunk.ReadInt64();
+                GridPeriod = chunk.ReadDouble();
+                GridOffset = chunk.ReadDouble();
                 Tempo = chunk.ReadSingle();
                 TimeSigNumBeatsBar = chunk.ReadByte();
                 TimeSigBeatValue = chunk.ReadByte();
